added ks test values
