+++ src/IndelCalling/ComparedSets.py
@@ -1,0 +1,3 @@
+from collections import namedtuple
+
+ComparedSets = namedtuple("ComparedSets", ['first_set', 'second_set'])
